backend/requirements.txt
flask==3.0.0
flask-cors==4.0.0
firebase-admin==6.6.0
google-cloud-firestore
google-cloud-storage
google-cloud-speech
google-cloud-aiplatform
python-dotenv==1.0.0
werkzeug==3.0.1
pydub==0.25.1
